% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIGD.R
\name{calcIGD}
\alias{calcIGD}
\title{Inverted Generational Distance}
\usage{
calcIGD(Y, Yref)
}
\arguments{
\item{Y}{Matrix of points in the objective space}

\item{Yref}{Matrix of Pareto-optimal reference points}
}
\value{
igd value (scalar)
}
\description{
Calculate IGD
}
