% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/check_stop_criteria.R
\name{check_stop_criteria}
\alias{check_stop_criteria}
\title{Stop criteria for MOEA/D}
\usage{
check_stop_criteria()
}
\value{
The function does not return any value. However, it modifies the
state of the calling environment, updating the iteration flag
\code{keep.running}.
}
\description{
Implements different stop criteria for the MOEA/D
}
\section{Parameters}{

This routine accesses all variables defined in the calling environment using
\code{parent.frame()}, so it does not require any explicit input parameters.
However, the calling environment must contain:
\itemize{
   \item \code{iters()} : counter function that registers the iteration
                          number
   \item \code{nfe()} : counter function that registers the number of
                          candidate solutions evaluated.
   \item \code{stopcrit}: list defining the stop criteria for the MOEA/D.
                          This list can contain the following fields:
   \itemize{
     \item \code{$maxiter}: positive integer indicating the maximum
                             number of iterations allowed.
     \item \code{$maxeval}: positive integer indicating the maximum
                             number of candidate solutions evaluated.
     \item \code{$names = c("stop_maxiter", "stop_maxeval")}:
           character vector containing the names of the
           stop criteria to be used.
   }
}
}

