% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraint_none.R
\name{constraint_none}
\alias{constraint_none}
\title{None constraint handling method for MOEA/D}
\usage{
constraint_none(B, bigZ, bigV, beta, ...)
}
\arguments{
\item{B}{Matrix of neighborhoods (generated by \code{define_neighborhood(...))})}

\item{bigZ}{Matrix of scalarized objective values for each neighborhood and the
incumbent solution (generated by \code{scalarize_values})}

\item{bigV}{Matrix of violation values for each neighborhood and the
incumbent solution}

\item{...}{other parameters (unused, included for compatibility with
generic call)}
}
\value{
[N x (T+1)] matrix of preference indexes. Each row contains
the T indexes of the candidate solutions in the neighborhood of
a given subproblem, plus a value of T+1 for the incumbent solution for
that subproblem.
}
\description{
Construct the preference index matrix based only on performance values.
}
\details{
This function ignores the violation values when constructing the preference index
matrix, using only the scalarized performance values.
}
\section{Parameters}{

}

