% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraint_tournament.R
\name{constraint_tournament}
\alias{constraint_tournament}
\title{"Tournament" constraint handling method for MOEA/D}
\usage{
constraint_tournament(B, bigZ, bigV, threshold, ...)
}
\arguments{
\item{B}{Matrix of neighborhoods (generated by \code{define_neighborhood(...))})}

\item{bigZ}{Matrix of scalarized objective values for each neighborhood and the
incumbent solution (generated by \code{scalarize_values})}

\item{bigV}{Matrix of violation values for each neighborhood and the
incumbent solution}

\item{threshold}{A real-valued, non-negative feasibility treshold}

\item{...}{other parameters (unused, included for compatibility with
generic call)}
}
\value{
\link{N x (T+1)} matrix of preference indexes. Each row contains
the T indexes of the candidate solutions in the neighborhood of
a given subproblem, plus a value of T+1 for the incumbent solution for
that subproblem, using the penalty-based constraint handling method.
}
\description{
Uses the Tournament Constraint handling method to generate a
preference index.
}
\details{
This function calculates the preference index of a set of neighborhoods
based on the "Tournament" constraint handling method. Please
see {\code{help(order_neighborhood)}} for more information on the
preference index matrix.

We define an individual as "Feasible", if its violation value V is below
a user defined Threshold. Given two individuals a_i and a_j in the same
neighborhood, they are ordered according to the following rule:

\itemize{
\item If both a_i and a_j are feasible, they are ordered by smallest performance value
\item If either a_i or a_j ar unfeasible, they are ordered by smallest violation value
}
}
\section{Parameters}{

}

