% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/decompose_problem.R
\name{decompose_problem}
\alias{decompose_problem}
\title{Calculate decomposition vectors}
\usage{
decompose_problem(decopars, m)
}
\arguments{
\item{decopars}{List of decomposition method parameters
See \code{Decomposition methods} for details.}

\item{m}{Number of objectives of the problem}
}
\value{
Matrix with one decomposition vector per row.
}
\description{
Calculates decomposition vectors for the MOEA/D.
}
\details{
This routine calculates the decomposition vectors for the MOEA/D. The
following methods for the generation of the weight vectors are currently
available:
\itemize{
   \item \code{$name = "Das"}: uses the method of Das and Dennis (1998). If
       this method is chosen, the function requires an additional field
       \code{decopars$H} containing a nonnegative integer value used to
       calculate the number of decomposition vectors,
       \eqn{n_{vecs} = \binom{(H + m - 1),(m - 1)}}
   \item \code{$name = "Li"}: uses the method of Li et al. (2014). If this
       method is chosen, the function requires additional fields
       \code{decopars$H} (similar to the one for the "Das" method), and
       \code{decopars$tau}, containing a contraction factor between 0 and 1.\cr
       If \code{m <= 6} this method is identical to "Das".\cr
       If \code{m > 6} the input \code{decopars$H} must be a 2-value vector.
       See Li et al. (2014) for details.
   \item \code{$name = "Uniform"}: Uniformly distributed vectors. If
       this method is chosen, the function requires an additional field
       \code{decopars$nvecs} containing the desired number of weight vectors.
}

Suggested values for \code{decopars$H} are (use with caution):

\tabular{cccc}{
\code{m}\tab \code{method}\tab \code{H}\tab n_{vecs}\cr

2    \tab "Das" \tab  \code{99}      \tab 100\cr
3    \tab "Das" \tab  \code{12}      \tab 91\cr
5    \tab "Das" \tab  \code{6}       \tab 210\cr
8    \tab "Li"  \tab  \code{c(3, 2)} \tab 156\cr
10   \tab "Li"  \tab  \code{c(3, 2)} \tab 275\cr
}

It is important to highlight that the number of vectors generated (based on
\code{decopars$H} for the "Das" and "Li" methods, and on
\code{decopars$nvecs} for the "Uniform" method) must be greater than the
number of neighbors declared in \code{decopars$neighbors}. This function will
throw an error if this condition is not verified.
}
\section{References}{

I. Das, J. Dennis (1998), "Normal Boundary Intersection - A New Method
for Generating the Pareto Surface in Nonlinear Multicriteria Optimization
Problems", SIAM J. Optim., 8(3), 631-657. \cr
K. Li et al. (2014), "An Evolutionary Many-Objective Optimization
Algorithm Based on Dominance and Decomposition", IEEE Trans. Evol. Comp.,
DOI: 10.1109/TEVC.2014.2373386
}

