% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomposition_msld.R
\name{decomposition_msld}
\alias{decomposition_msld}
\title{Problem Decomposition using Multi-layered Simplex-lattice Design}
\usage{
decomposition_msld(decomp)
}
\arguments{
\item{decomp}{TODO}
}
\description{
Problem Decomposition using Multi-layered Simplex-lattice Design for MOEADr
package
}
\details{
This routine calculates the weight vectors for the MOEA/D using the
Multi-layered Simplex-lattice Design.
}
\section{References}{

K. Li et al. (2014), "An Evolutionary Many-Objective Optimization
Algorithm Based on Dominance and Decomposition",
IEEE Trans. Evol. Comp. 19(5):694-716, 2015. DOI: 10.1109/TEVC.2014.2373386
}

