% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_neighborhood.R
\name{define_neighborhood}
\alias{define_neighborhood}
\title{Calculate neighborhood relations}
\usage{
define_neighborhood(neighbors = NULL, v.matrix = NULL)
}
\arguments{
\item{neighbors}{List containing the decomposition method parameters.
If \code{NULL} the function searches for \code{neighbors} in the calling
environment.}

\item{v.matrix}{matrix of vectors for defining neighborhoods. If NULL it is
captured directly from the calling environment, depending on
\code{neighbors$name}).}
}
\value{
List containing the matrix of selection probabilities (\code{P}) and
the matrix of neighborhoods (\code{B}).
}
\description{
Calculates neighborhood relations for the MOEADr package
}
\details{
This routine calculates the neighborhood relations for the MOEA/D. The
following methods are currently available:

#' \itemize{
\item \code{neighbors$name = "lambda"}: defines the neighborhood using the
distance matrix for the weight vectors. The calculation is performed only
once for the entire run.
\item \code{neighbors$name = "x"}: defines the neighborhood using
the distance matrix for the incumbent solution associated with each
subproblem. In this case the calculation is performed at each iteration.
}

The number of neighbors declared in \code{neighbors$T} must be smaller than
the number of subproblems (defined in \code{generate_weights}).

The probability of sampling from the neighborhood when performing variation,
\code{neightbors$delta.p}, must be a scalar value between 0 and 1.
}
