% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DEops.R
\name{DEops}
\alias{DEops}
\title{Differential Mutation operators}
\usage{
DEops()
}
\value{
Matrix \code{U} containing the resulting population
}
\description{
Implementation of the Differential Evolution variation operators for the
MOEA/D
}
\details{
This function performs the variation operators of the DE algorithm. For
each candidate solution \eqn{X[i, ]}, a candidate vector \eqn{V[i, ]}
is generated according to:

\code{V [i, ] = X [i, ]}

\code{if runif(1) < CR, V [i, j] = X [i, j] + F * (X [r1, j] + X [r2, j]),
     with j = 1, ..., n}

wherein \code{r1 != r2 != i} are mutually exclusive, randomly
selected indices that indicate vectors from either the neighborhood of
\code{i} (with probability \code{delta}) or the whole population \code{X}
(with probability \code{1 - delta})
}
\section{Parameters}{

This routine accesses all variables defined in the calling environment using
\code{parent.frame()}, so it does not require any explicit input parameters.
However, the calling environment must contain:
\itemize{
   \item a population matrix \code{X}
   \item a neighborhood matrix \code{N}
   \item a list of variation parameters \code{chngpars}, containing:
   \itemize{
       \item \code{chngpars$DE_diffmut$F} : scaling factor (REQUIRED).
       \item \code{chngpars$DE_diffmut$CR} : recombination probability
           (OPTIONAL. Defaults to 1).
       \item \code{chngpars$DE_diffmut$delta} : probability of selecting the
           'parent' vectors from the neighborhood of each candidate solution.
           (OPTIONAL. Defaults to 0.9).
   }
}
}

\section{References}{

K. Price, R.M. Storn, J.A. Lampinen, "Differential Evolution: A
Practical Approach to Global Optimization", Springer 2005.\cr

H. Li, Q. Zhang, "Multiobjective Optimization Problems With Complicated
Pareto Sets, MOEA/D and NSGA-II", IEEE. Trans. Evol. Comp. 12(2):284-302,
2009.
}

