% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_problem.R
\name{example_problem}
\alias{example_problem}
\title{Example problem}
\usage{
example_problem(X)
}
\arguments{
\item{X}{population matrix (see \code{\link[=moead]{moead()}} for details)}
}
\value{
Matrix of objective function values
}
\description{
Example problem - minimization of shifted sphere and rastrigin functions.
}
