% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_constraint_methods.R
\name{get_constraint_methods}
\alias{get_constraint_methods}
\title{Print available constraint methods}
\usage{
get_constraint_methods()
}
\value{
Formatted data frame containing reference name (for
\code{constraint$name}) and instructions for More Info about each method.
}
\description{
Prints the constraint handling methods available in the MOEADr package
}
\details{
This routine prints the names of the constraint handling methods available in
the MOEADr package, to be used as the \code{constraint$name} parameter in the
\code{moead(...)} call. Instructions for obtaining more info on each
operator are also returned.
}
