% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_decomposition_methods.R
\name{get_decomposition_methods}
\alias{get_decomposition_methods}
\title{Print available decomposition methods}
\usage{
get_decomposition_methods()
}
\value{
Formatted data frame containing reference name (for
\code{decomp$name}) and instructions for More Info about each method.
}
\description{
Prints the decomposition methods available in the MOEADr package
}
\details{
This routine prints the names of the decomposition methods available in
the MOEADr package, to be used as the \code{decomp$name} parameter in the
\code{moead(...)} call. Instructions for obtaining more info on each
operator are also returned.
}
