% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_localsearch_methods.R
\name{get_localsearch_methods}
\alias{get_localsearch_methods}
\title{Print available local search methods}
\usage{
get_localsearch_methods()
}
\value{
Formatted data frame containing reference name (for
\code{variation$localsearch$type}) and instructions for More Info about
each method.
}
\description{
Prints the local search methods available in the MOEADr package
}
\details{
This routine prints the names of the local search methods available in
the MOEADr package, to be used as the \code{aggfun$name} parameter in the
\code{moead(...)} call. Instructions for obtaining more info on each
operator are also returned.
}
