% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_repair_operators.R
\name{get_repair_operators}
\alias{get_repair_operators}
\title{Print available repair operators}
\usage{
get_repair_operators()
}
\value{
Formatted data frame containing reference name (for
\code{repair$name}) and instructions for More Info about each operator.
}
\description{
Prints the repair operators available in the MOEADr package
}
\details{
This routine prints the names of the repair operators available in
the MOEADr package, to be used as the \code{repair$name} parameter in the
\code{moead(...)} call. Instructions for obtaining more info on each
operator are also returned.
}

