% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scalarization_methods.R
\name{get_scalarization_methods}
\alias{get_scalarization_methods}
\title{Print available scalarization methods}
\usage{
get_scalarization_methods()
}
\value{
Formatted data frame containing reference name (for
\code{aggfun$name}) and instructions for More Info about each method.
}
\description{
Prints the scalarization methods available in the MOEADr package
}
\details{
This routine prints the names of the scalarization methods available in
the MOEADr package, to be used as the \code{aggfun$name} parameter in the
\code{moead(...)} call. Instructions for obtaining more info on each
operator are also returned.
}
