% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stop_criteria.R
\name{get_stop_criteria}
\alias{get_stop_criteria}
\title{Print available stop criteria}
\usage{
get_stop_criteria()
}
\value{
Formatted data frame containing reference name (for
\code{stopcrit[[i]]$name}) and instructions for More Info about each
criterion.
}
\description{
Prints the stop criteria available in the MOEADr package
}
\details{
This routine prints the names of the stop criteria available in
the MOEADr package, to be used as the \code{stopcrit[[i]]$name} parameter
in the \code{moead(...)} call. Instructions for obtaining more info on each
criterion are also returned.
}
