% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_update_methods.R
\name{get_update_methods}
\alias{get_update_methods}
\title{Print available update methods}
\usage{
get_update_methods()
}
\value{
Formatted data frame containing reference name (for
\code{update$name}) and instructions for More Info about each method.
}
\description{
Prints the update methods available in the MOEADr package
}
\details{
This routine prints the names of the update methods available in
the MOEADr package, to be used as the \code{update$name} parameter in the
\code{moead(...)} call. Instructions for obtaining more info on each
operator are also returned.
}
