% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_variation_operators.R
\name{get_variation_operators}
\alias{get_variation_operators}
\title{Print available variation operators}
\usage{
get_variation_operators()
}
\value{
Formatted data frame containing reference name (for
\code{variation$name}) and instructions for More Info about each operator.
}
\description{
Prints the variation operators available in the MOEADr package
}
\details{
This routine prints the names of the variation operators available in
the MOEADr package, to be used as the \code{variation$name} parameter in the
\code{moead(...)} call. Instructions for obtaining more info on each
operator are also returned.
}
