% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadPlotData.R
\name{loadPlotData}
\alias{loadPlotData}
\title{loadPlotData}
\usage{
loadPlotData(name, j, wd, extra = T)
}
\arguments{
\item{name}{filename to load}

\item{wd}{which direction to load}

\item{moead}{moead class output}
}
\value{
nothing
}
\description{
loadPlotData
}
\details{
load moead Data
}
\section{References}{


F. Campelo, L.S. Batista, C. Aranha (2020): The {MOEADr} Package: A
Component-Based Framework for Multiobjective Evolutionary Algorithms Based on
Decomposition. Journal of Statistical Software \doi{10.18637/jss.v092.i06}\cr
}

