% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_dvls.R
\name{ls_dvls}
\alias{ls_dvls}
\title{Differential vector-based local search}
\usage{
ls_dvls(X, Y, B, W, which.x, eps = 1e-06, ...)
}
\arguments{
\item{X}{Population matrix}

\item{Y}{Matrix of objective function values for X}

\item{B}{Matrix of neighborhoods (generated by \code{\link[=define_neighborhood]{define_neighborhood()}})}

\item{W}{matrix of weights (generated by \code{\link[=generate_weights]{generate_weights()}}).}

\item{which.x}{logical vector indicating which subproblems should undergo
local search}

\item{eps}{threshold for using the quadratic approximation value}

\item{...}{other parameters (unused, included for compatibility with
generic call)}
}
\value{
Matrix \code{X}' containing the modified population
}
\description{
Differential vector-based local search (DVLS) implementation for the MOEA/D
}
\details{
This routine implements the differential vector-based local search for
the MOEADr package. Check the references for details.

This routine is intended to be used internally by \code{\link[=variation_localsearch]{variation_localsearch()}},
and should not be called directly by the user.
}
\section{References}{

B. Chen, W. Zeng, Y. Lin, D. Zhang,
"A new local search-based multiobjective optimization algorithm",
IEEE Trans. Evolutionary Computation 19(1):50-73, 2015.
}

