% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_tpqa.R
\name{ls_tpqa}
\alias{ls_tpqa}
\title{Three-point quadratic approximation local search}
\usage{
ls_tpqa(X, Y, B, W, which.x, eps = 1e-06, ...)
}
\arguments{
\item{X}{Population matrix}

\item{Y}{Matrix of objective function values for X}

\item{B}{Matrix of neighborhoods (generated by \code{\link[=define_neighborhood]{define_neighborhood()}})}

\item{W}{matrix of weights (generated by \code{\link[=generate_weights]{generate_weights()}}).}

\item{which.x}{logical vector indicating which subproblems should undergo
local search}

\item{eps}{threshold for using the quadratic approximation value}

\item{...}{other parameters (unused, included for compatibility with
generic call)}
}
\value{
Matrix \code{X}' containing the modified population
}
\description{
Three-point quadratic approximation (TPQA) local search implementation for
the MOEA/D
}
\details{
This routine implements the 3-point quadratic approximation local search for
the MOEADr package. Check the references for details.
}
\section{References}{

Y. Tan, Y. Jiao, H. Li, X. Wang,
"A modification to MOEA/D-DE for multiobjective optimization problems with
complicated Pareto sets",
Information Sciences 213(1):14-38, 2012.
}

