% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/moead.R
\name{moead}
\alias{moead}
\title{MOEA/D}
\usage{
moead(probpars, decopars, scalpars, chngpars, updtpars, showpars, stopcrit,
  seed = NULL)
}
\arguments{
\item{probpars}{List of problem parameters.
See \code{Problem Description} for details.}

\item{decopars}{List of decomposition method parameters
See \code{Decomposition methods} for details.}

\item{scalpars}{List of scalarization method parameters
See \code{Scalarization methods} for details.}

\item{chngpars}{List of variation operator parameters
See \code{Variation operators} for details.}

\item{stopcrit}{list of stop criteria parameters. See
   \code{Stop criteria} for details.}
}
\description{
MOEA/D implementation in R
}
\details{
MOEA/D implementation in R
}
\section{Problem Description}{

The \code{probpars} parameter consists of a list with all necessary
definitions for the multiobjective optimization problem to be solved.
\code{probpars} must contain at least the following fields:
\itemize{
   \item \code{$name} - name of the problem instance function, that is, a
         routine that calculates y = f(x);
   \item \code{$xmin} - vector of lower bounds of each variable
   \item \code{$xmax} - vector of upper bounds of each variable
   \item \code{$nobj} - integer containing the number of objectives
}

The function indicated in \code{probpars$name} must be able to receive a
matrix with each row representing one candidate solution, and return a matrix
with each row representing the objective values for that solution.
}

\section{Decomposition Methods}{

The \code{decopars} parameter defines the method to be used for the
generation of the decomposition vectors. \code{decopars} must have
at least the \code{$name} parameter. Currently implemented methods are:

\itemize{
   \item \code{$name = "Das"}      - Das and Dennis (1998);
   \item \code{$name = "Li"}       - Li et al. (2014);
   \item \code{$name = "Uniform"}  - Uniformly distributed vectors;
}

Check \code{\link{decompose_problem}} for details.
}

\section{Scalarization Methods}{

The \code{scalpars} parameter defines the method to be used for the
scalarization of the multiobjective function. \code{scalpars} must have
at least the \code{$name} parameter. Currently implemented methods are:
\itemize{
   \item \code{$name = "ws"}  - Weighted Sum;
   \item \code{$name = "wt"}  - Weighted Tchebycheff;
   \item \code{$name = "pbi"} - Penalty-based Boundary Intersection;
   \item \code{$name = "mwt"} - Modified Weighted Tchebycheff;
}

See \code{\link{scalarize_values}} for details.
}

\section{Variation Operators}{

TODO: Write a description of the \code{chngpars} structure.
}

\section{Stop Criteria}{

TODO: Write a description of the \code{stopcrit} structure.
}

\section{References}{

Q. Zhang and H. Li, "MOEA/D: A Multiobjective Evolutionary Algorithm
}
\examples{
# MOEA/D as in Zhang and Li (2007) (sec. V-E, p.721-722)
library(GPareto) # For the test problems
probpars <- list(name = "DTLZ", prob.number = 1, nobj = 3,
                 xmin = rep(0, 10), xmax = rep(1, 10))
decopars <- list(name = "Das", H = 25, neighbors = 20)
scalpars <- list(name = "wt")
chngpars <- list(sbx     = list(eta = 20, pc = 1, eps = 1e-16),
                 polymut = list(eta = 20, pm = 0.1))
updtpars <- list(method = "standard")
showpars <- list(show.iters = "numbers", showevery = 10)
stopcrit <- list(names = "stop_maxiter",
                 maxiter = 250)
seed = NULL
OUT1 <- moead(probpars, decopars, scalpars,
              chngpars, updtpars, showpars,
              stopcrit, seed)

# MOEA/D as in Zhang and Li (2007) (using PBI, sec. V-F, p. 724)
scalpars <- list(name = "pbi", theta = 5)
OUT2 <- moead(probpars, decopars, scalpars,
              chngpars, updtpars, showpars,
              stopcrit, seed)

# MOEA/D as in Zhang and Li (2007) (with obj normalization, sec. V-F, p. 724)
scalpars <- list(name = "wt", normalize.obj = TRUE)
probpars <- probpars <- list(name = "ZDT", prob.number = 1, nobj = 2,
                 xmin = rep(0, 30), xmax = rep(1, 30))
chngpars <- list(sbx     = list(eta = 20, pc = 1, eps = 1e-16),
                 polymut = list(eta = 20, pm = 1/30))
decopars <- list(name = "Das", H = 149, neighbors = 20)
OUT3 <- moead(probpars, decopars, scalpars,
              chngpars, updtpars, showpars,
              stopcrit, seed)

#' # MOEA/D-DE as in Zhang and Li (2009)
scalpars <- list(name = "wt", normalize.obj = FALSE)
probpars <- probpars <- list(name = "ZDT", prob.number = 1, nobj = 2,
                 xmin = rep(0, 30), xmax = rep(1, 30))
chngpars <- list(DEops   = list(delta = 0.9, CR = 1, F = 0.5),
                 polymut = list(eta = 20, pm = 1/30))
decopars <- list(name = "Das", H = 299, neighbors = 20)
stopcrit <- list(names = "stop_maxiter",
                 maxiter = 500)
OUT4 <- moead(probpars, decopars, scalpars,
              chngpars, updtpars, showpars,
              stopcrit, seed)
}

