% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_neigborhood.R
\name{order_neighborhood}
\alias{order_neighborhood}
\title{Order Neighborhood for MOEA/D}
\usage{
order_neighborhood(moead.env)
}
\arguments{
\item{moead.env}{list representing the environment of the base function
\code{moead}.}
}
\value{
[N x (T+1)] matrix of preference indexes. Each row contains
the T indexes of the candidate solutions in the neighborhood of
a given subproblem, plus a value of T+1 for the incumbent solution for
that subproblem, in an order defined by the constraint handling
method defined.
}
\description{
Calculates the order of a candidate matrix based on their scalarized performance and violation values
}
\details{
This routine receives a matrix of scalarized performance values, a neighborhood matrix,
and a list of violation values, and calculates the preference order of the candidates
for each neighborhood based on the performance values and the constraint handling method.
}
\section{Parameters}{

This routine receives one input variable:
\code{moead.env} is generated within the calling function \code{update_population()}.
See \code{\link{update_population}} for more information.

The environment listed in \code{moead.env} must contain the variable
\code{constraint} which provides the type of constraint handling
to be employed and the relevant parameters. The list of available
constraint handling methods can be generated using
\code{get_constraint_methods()}
}

