% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_variation.R
\name{perform_variation}
\alias{perform_variation}
\title{Run variation operators}
\usage{
perform_variation(X, P, B, W, variation, ...)
}
\arguments{
\item{X}{Population matrix of the MOEA/D (each row is a candidate solution).}

\item{P}{Matrix of probabilities of selection for variation (created by
\code{\link[=define_neighborhood]{define_neighborhood()}}).}

\item{B}{Matrix of neighborhood indexes (created by \code{\link[=define_neighborhood]{define_neighborhood()}}).}

\item{W}{matrix of weights (created by \code{\link[=generate_weights]{generate_weights()}}).}

\item{variation}{List vector containing the variation operators to be used.
See \code{\link[=moead]{moead()}} for details.}

\item{...}{other parameters (included for compatibility with generic call)}
}
\value{
List object containing a modified population matrix \code{X} and a
local search argument list \code{ls.arg}.
}
\description{
Sequentially apply variation operators for the MOEADr package
}
\details{
This routine performs the variation block for the MOEA/D. The
list of available variation operators can be generated using
\code{\link[=get_variation_operators]{get_variation_operators()}}.

If the \code{localsearch} operator is included, it is executed whenever its
conditions (period of occurrence or probability of occurrence) are verified.
See \code{\link[=variation_localsearch]{variation_localsearch()}} for details.

\strong{Warning}: this routine (and the ones called within it) may access (but
not directly modify) variables from the calling environment, depending on
the specific operators being performed.
}
