% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_moeadoutput.R
\name{plot.moeadoutput}
\alias{plot.moeadoutput}
\title{plot.moeadoutput}
\usage{
\method{plot}{moeadoutput}(output, d = c(1, 2), ...)
}
\description{
S3 method for plotting the output of an moead run
}
\details{
This function plots the pareto front of the solution set found by moead.
By default, the first and second objectives are plotted, but other objectives
can be specified as parameters.
}
