% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/polymut.R
\name{polymut}
\alias{polymut}
\title{Polynomial mutation}
\usage{
polymut()
}
\value{
Matrix \code{U} containing the recombined population
}
\description{
Polynomial mutation implementation for the MOEA/D
}
\details{
This R implementation of the Polynomial Mutation reproduces the C code
implementation available in package "emoa", version 0.5-0, by Olaf Mersmann.
The differences between the present version and the original one are:
\itemize{
   \item All variables are considered bounded in the (0, 1) interval, which
         simplifies some calculations.
   \item Calculations are vectorized over variables, which also simplifies
         some operations.
}
}
\section{Parameters}{

This routine accesses all variables defined in the calling environment using
\code{parent.frame()}, so it does not require any explicit input parameters.
However, the calling environment must contain:
\itemize{
   \item a population matrix \code{X}
   \item a list of variation parameters \code{chngpars}, containing:
   \itemize{
       \item \code{chngpars$polymut$eta} : mutation parameter (REQUIRED)
       \item \code{chngpars$polymut$pm}  : variable-wise probability of
                                     mutation (REQUIRED)
   }
}
}

\section{References}{

K. Deb and S. Agrawal (1999). A Niched-Penalty Approach for Constraint
Handling in Genetic Algorithms. In: Artificial Neural Nets and Genetic
Algorithms, pp. 235-243, Springer.

Olaf Mersmann (2012). emoa: Evolutionary Multiobjective
Optimization Algorithms. R package version 0.5-0.\cr
http://CRAN.R-project.org/package=emoa
}

