% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preset_moead.R
\name{preset_moead}
\alias{preset_moead}
\title{preset_moead}
\usage{
preset_moead(name = NULL)               -- returns the list of available
                                           preset names
preset_moead(name = "preset_name")      -- returns the specific preset as a
                                           configuration list
}
\description{
Library of configuration presets for use in the MOEADr package.
}
\details{
This function contains a list of configuration presets taken from
the literature to be used with the \link{moead} function in the moeadr
package.

Use these configurations as a starting point, but we strongly
recommend that you play around with the particular configurations
(see example).
}
\examples{
configuration <- preset_moead(name = "original")
configuration$stopcrit = list(list(name = "maxiter", maxiter = 50))

library(smoof)
ZDT1 <- make_vectorized_smoof(prob.name  = "ZDT1",
                              dimensions = 30)
                              problem   <- list(name       = "ZDT1",
                                                xmin       = rep(0, 30),
                                                xmax       = rep(1, 30),
                                                m          = 2)
showpars  <- list(show.iters = "dots", showevery  = 10)
seed      <- 42

output <- moead(problem = problem,
                preset = configuration,
                showpars = showpars, seed = seed)

}
