% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repair_none.R
\name{repair_none}
\alias{repair_none}
\title{Identity operator}
\usage{
repair_none(X, ...)
}
\arguments{
\item{X}{Population matrix}

\item{...}{other parameters (unused, included for compatibility with
generic call)}
}
\value{
Matrix \code{X}
}
\description{
Identity operator (no repair performed)
}
\details{
Performs the identity operator (no repair). This routine is included to
simplify the use of automated tuning / design tools such as Iterated Racing.
}

