% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repair_truncate.R
\name{repair_truncate}
\alias{repair_truncate}
\title{Truncate Repair}
\usage{
repair_truncate(X, ...)
}
\arguments{
\item{X}{Population matrix}

\item{...}{other parameters (unused, included for compatibility with
generic call)}
}
\value{
Matrix \code{X}
}
\description{
Truncation repair operation
}
\details{
Repairs the solution matrix X to guarantee that all values are in the
[0,1] interval. This is done by simply truncating the matrix to 0,1.

Code copied shamelessly from fcampelo's ExpDE repository.
https://github.com/fcampelo/ExpDE/blob/master/R/ExpDE.R
Lines 240-242
}

