% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_allocation_select.R
\name{resource_allocation_select}
\alias{resource_allocation_select}
\title{Resource Allocation select}
\usage{
resource_allocation_select(
  iter,
  resource.allocation,
  W,
  priority.values,
  problem,
  idx.boundary = NULL,
  ...
)
}
\arguments{
\item{iter}{integer iteration}

\item{resource.allocation}{list resource allocation method and specific vars}

\item{W}{matrix of weights}

\item{priority.values}{priority values for deciding which subproblem to give resources}

\item{idx.boundary}{indexes of the boundaries subproblem}

\item{...}{other parameters (included for compatibility with generic call)}
}
\value{
indexes and counter of subproblem usage \code{out}'.
}
\description{
Proxy for Resource Allocation selecting procedure
}
\details{
Returns a vector with indexes of each subproblem that will receive resource
}
\section{References}{


F. Campelo, L.S. Batista, C. Aranha (2020): The {MOEADr} Package: A
Component-Based Framework for Multiobjective Evolutionary Algorithms Based on
Decomposition. Journal of Statistical Software \doi{10.18637/jss.v092.i06}\cr
}

