% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_allocation_select_none.R
\name{resource_allocation_select_none}
\alias{resource_allocation_select_none}
\title{Resource Allocation dummy, nothing happens here}
\usage{
resource_allocation_select_none(
  iter,
  resource.allocation,
  W,
  priority.values,
  problem,
  ...
)
}
\arguments{
\item{resource.allocation}{which Resource Allocation method}

\item{W}{matrix of weight}

\item{...}{other parameters (included for compatibility with generic call)}

\item{bigZ}{bigZ legado}
}
\value{
list of indexes and use counter for subproblems \code{out}'.
}
\description{
Resource Allocation None - dummy
}
\section{References}{


F. Campelo, L.S. Batista, C. Aranha (2020): The {MOEADr} Package: A
Component-Based Framework for Multiobjective Evolutionary Algorithms Based on
Decomposition. Journal of Statistical Software \doi{10.18637/jss.v092.i06}\cr
}

