% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savePlotData.R
\name{savePlotData}
\alias{savePlotData}
\title{savePlotData}
\usage{
savePlotData(moea, name, j, wd = "~/dataExp/", extra = T)
}
\arguments{
\item{name}{filename to save}

\item{wd}{which direction to save}

\item{moead}{moead class output}
}
\value{
nothing
}
\description{
savePlotData
}
\details{
save moead Data
}
\section{References}{


F. Campelo, L.S. Batista, C. Aranha (2020): The {MOEADr} Package: A
Component-Based Framework for Multiobjective Evolutionary Algorithms Based on
Decomposition. Journal of Statistical Software \doi{10.18637/jss.v092.i06}\cr
}

