% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sbx.R
\name{sbx}
\alias{sbx}
\title{Simulated binary crossover}
\usage{
sbx()
}
\value{
Matrix \code{U} containing the recombined population
}
\description{
SBX implementation for the MOEA/D
}
\details{
This R implementation of the Simulated Binary Crossover reproduces the C code
implementation available in package "emoa", version 0.5-0, by Olaf Mersmann.
The differences between the present version and the original one are:
\itemize{
   \item All variables are considered bounded in the (0, 1) interval, which
         simplifies some calculations.
   \item Calculations are vectorized over variables, which also simplifies
         some operations.
}
}
\section{Parameters}{

This routine accesses all variables defined in the calling environment using
\code{parent.frame()}, so it does not require any explicit input parameters.
However, the calling environment must contain:
\itemize{
   \item a population matrix \code{X}
   \item a neighborhood matrix \code{N}
   \item a list of variation parameters \code{chngpars}, containing:
   \itemize{
       \item \code{chngpars$sbx$eta} : spread constant (REQUIRED)
       \item \code{chngpars$sbx$pc}  : variable-wise probability of crossover
                           (optional, defaults to 1)
       \item \code{chngpars$sbx$eps} : smallest difference considered by
                           crossover (optional, defaults to 1e-12)
   }
}
}

\section{References}{

Deb, K. and Agrawal, R. B. (1995) Simulated binary crossover for continuous
search space. Complex Systems, 9 115-148 \cr

Olaf Mersmann (2012). emoa: Evolutionary Multiobjective
Optimization Algorithms. R package version 0.5-0.\cr
http://CRAN.R-project.org/package=emoa
}

