% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scalarize_values.R
\name{scalarize_values}
\alias{scalarize_values}
\title{Scalarize values for MOEA/D}
\usage{
scalarize_values(Y, W, scalpars, minP = NULL, maxP = NULL)
}
\arguments{
\item{Y}{matrix containing the objective function values to be scalarized
(each row corresponds to a candidate solution)}

\item{W}{matrix of weights. Must have the same dimensions as Y.}

\item{scalpars}{List of scalarization method parameters
See \code{Scalarization methods} for details.}

\item{minP}{current estimate of the 'ideal' point. Defaults to NULL, in
which case the point is estimated from \code{Y}.}

\item{maxP}{current estimate of the 'nadir' point. Defaults to NULL, in
            which case the point is estimated from \code{Y}.}
}
\value{
Vector of scalarized values.
}
\description{
Perform scalarization for the MOEA/D.
}

