% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_objectives.R
\name{scale_objectives}
\alias{scale_objectives}
\title{Scaling of the objective function values}
\usage{
scale_objectives(moead.env, eps = 1e-16)
}
\arguments{
\item{moead.env}{list representing the environment of the base function
\code{moead}.}

\item{eps}{tolerance value for avoiding divisions by zero.}
}
\value{
List object containing scaled objective function value matrices
\code{Y} and \code{Yt}, as well as "ideal" point \code{minP} and "nadir"
point \code{maxP}.
}
\description{
Performs scaling of the objective function values for the MOEADr package
}
\details{
This routine scales the matrices of objective function values for the
current (\code{Yt}) and candidate (\code{Y}) solutions. The
following methods are currently available:

\itemize{
   \item \code{scaling$name = "none"}: no scaling
   \item \code{scaling$name = "simple"}: simple linear scaling between
   estimated ideal and nadir points, calculated from the available points in
   \code{Y} and \code{Yt} at each iteration.
}
}
\section{Parameters}{

This routine receives a single input variable, \code{moead.env}, which is
generated within the calling function \code{update_population()}
(see \code{\link{update_population}} for more information).
This environment must contain at least the following variables:

\itemize{
   \item \code{Yt}: matrix of objective values for population \code{Xt}
   \item \code{Y} : matrix of objective values for population \code{X}
   \item \code{scaling}: list of function scaling parameters, containing at
   least the field \code{scaling$name}.
}
}

