% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop_maxeval.R
\name{stop_maxeval}
\alias{stop_maxeval}
\title{Stop criterion: maximum number of evaluations}
\usage{
stop_maxeval(moead.env)
}
\arguments{
\item{moead.env}{list representing the environment of the base function
\code{moead}.}
}
\value{
boolean value: TRUE if this criterion has been met, FALSE otherwise.
}
\description{
Verifies stop criterion "maximum number of evaluations" for the MOEADr
package. For internal use only, not to be called directly by the user.
}
\details{
When this stop criterion is used, one element of the \code{stopcrit}
parameter (see \code{\link{moead}}) must have the following structure:
\itemize{
\item \code{stopcrit$name = "maxeval"}
\item \code{stopcrit$maxeval = K}, where \code{K} is a positive integer.
}
}
