% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop_maxiter.R
\name{stop_maxiter}
\alias{stop_maxiter}
\title{Stop criterion: maximum number of iterations}
\usage{
stop_maxiter(moead.env)
}
\arguments{
\item{moead.env}{list representing the environment of the base function
\code{moead}.}
}
\value{
boolean value: TRUE if this criterion has been met, FALSE otherwise.
}
\description{
Verifies stop criterion "maximum number of iterations" for the MOEADr
package. For internal use only, not to be called directly by the user.
}
\details{
When this stop criterion is used, one element of the \code{stopcrit}
parameter (see \code{\link{moead}}) must have the following structure:
\itemize{
   \item \code{stopcrit$name = "maxiter"}
   \item \code{stopcrit$maxiter = K}, where \code{K} is a positive integer.
}
}

