% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop_maxtime.R
\name{stop_maxtime}
\alias{stop_maxtime}
\title{Stop criterion: maximum runtime}
\usage{
stop_maxtime(moead.env)
}
\arguments{
\item{moead.env}{list representing the environment of the base function
\code{moead}.}
}
\value{
boolean value: TRUE if this criterion has been met, FALSE otherwise.
}
\description{
Verifies stop criterion "run time limit" for the MOEADr
package. For internal use only, not to be called directly by the user.
}
\details{
When this stop criterion is used, one element of the \code{stopcrit}
parameter (see \code{\link{moead}}) must have the following structure:
\itemize{
\item \code{stopcrit$name = "maxtime"}
\item \code{stopcrit$maxtime = T}, where \code{T} is a positive integer
representing the desired time limit (in seconds).
}
}
\section{Warning}{

This function uses Sys.time() for verifying the total run time, i.e., it
counts wall-clock time, not CPU time.
}

