% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_moeadoutput.R
\name{summary.moeadoutput}
\alias{summary.moeadoutput}
\title{summary.moeadoutput}
\usage{
\method{summary}{moeadoutput}(output, ...)
}
\description{
S3 method for summarizing the output of an moead run
}
