% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/update_population.R
\name{update_population}
\alias{update_population}
\title{Update population}
\usage{
update_population()
}
\value{
The function does not return any value. However, it modifies the
state of the calling environment, updating the population matrix \code{X} and
the matrix of objective values \code{Y}.
}
\description{
Selection and population update procedures for the MOEA/D
}
\section{Parameters}{

This routine accesses all variables defined in the calling environment using
\code{parent.frame()}, so it does not require any explicit input parameters.
However, the calling environment must contain:
\itemize{
   \item \code{Xt}: current population matrix
   \item \code{Yt}: matrix of objective values for population \code{Xt}
   \item \code{X} : candidate population matrix
   \item \code{Y} : matrix of objective values for population \code{X}
   \item \code{N} : neighborhood matrix
   \item \code{W} : matrix of weigths
   \item \code{updtpars}: list of update parameters.
   \item \code{scalpars}: list of scalarization parameters.
}
}

\section{References}{

Q. Zhang and H. Li, "MOEA/D: A Multiobjective Evolutionary Algorithm
}

