% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updt_best.R
\name{updt_best}
\alias{updt_best}
\title{Best Neighborhood Replacement Update for MOEA/D}
\usage{
updt_best(call.env)
}
\arguments{
\item{call.env}{list representing the environment of the base function
\code{\link[=moead]{moead()}}. Variable \code{call.env$update} must have the following key-value
pairs:
\itemize{
\item \code{update$Tr}: positive integer, neighborhood size for the update
operation
\item \code{update$nr}: positive integer, maximum number of copies of a given
candidate solution.
}}
}
\value{
List object containing the update population matrix (\code{X}),
and its corresponding matrix of objective function values (\code{Y}) and
constraint value list (\code{V}).
}
\description{
Population update using the best neighborhood replacement method for the
MOEADr package.
}
\details{
The Best Neighborhood Replacement method consists of three steps:

\itemize{
\item For each subproblem \code{i}, the best candidate solution \code{x_j} from the
entire population is determined.
\item The neighborhood of subproblem \code{i} is replaced by the neighborhood
of subproblem j. The size of this neighborhood is given by a parameter
\code{Tr}.
\item The Restricted replacement (see \code{\link[=updt_restricted]{updt_restricted()}}) is then
applied using this new neighborhood.
}

This update routine is intended to be used internally by the main \code{\link[=moead]{moead()}}
function, and should not be called directly by the user.
}
