% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updt_standard.R
\name{updt_standard}
\alias{updt_standard}
\title{Standard Neighborhood Replacement Update for MOEA/D}
\usage{
updt_standard(call.env)
}
\arguments{
\item{call.env}{list representing the environment of the base function
\code{\link[=moead]{moead()}}.}
}
\value{
List object containing the update population matrix (\code{X}),
and its corresponding matrix of objective function values (\code{Y}) and
constraint value list (\code{V}).
}
\description{
Population update using the standard neighborhood replacement method for the
MOEADr package.
}
\details{
This routine executes the standard neighborhood replacement operation to
update the population matrix of the MOEA/D.
This update routine is intended to be used internally by the main \code{\link[=moead]{moead()}}
function, and should not be called directly by the user.
}
