% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variation_binrec.R
\name{variation_binrec}
\alias{variation_binrec}
\title{Binomial Recombination}
\usage{
variation_binrec(X, Xt, rho, ...)
}
\arguments{
\item{X}{Population matrix}

\item{Xt}{Original population matrix}

\item{rho}{mutation probability}

\item{...}{other parameters (included for compatibility with generic call)}
}
\value{
Matrix \code{X}' containing the recombined population
}
\description{
Binomial recombination implementation for the MOEA/D.
}
\details{
This variation operator only works if at least one other variation operator
is performed prior to its execution, otherwise it becomes an identity
operator (returns an unchanged matrix X).
}
\section{References}{

K. Price, R.M. Storn, J.A. Lampinen, "Differential Evolution: A
Practical Approach to Global Optimization", Springer 2005
}

