% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variation_binrecomb.R
\name{variation_binrecomb}
\alias{variation_binrecomb}
\title{Binomial Recombination}
\usage{
variation_binrecomb(X, Xc, rho, ...)
}
\arguments{
\item{X}{Population matrix}

\item{Xc}{Original Population matrix}

\item{rho}{mutation probability}

\item{...}{other parameters (unused, included for compatibility with
generic call)}
}
\value{
Matrix \code{X}' containing the mutated population
}
\description{
Binomial Recombination
}

