% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variation_diffmut.R
\name{variation_diffmut}
\alias{variation_diffmut}
\title{Differential Mutation}
\usage{
variation_diffmut(X = NULL, P = NULL, B = NULL, phi = NULL,
  basis = "random", ...)
}
\arguments{
\item{X}{Population matrix}

\item{P}{Matrix of selection probabilities (generated by
\code{define_neighborhood(...)})}

\item{phi}{Mutation parameter, random if NULL}

\item{basis}{how to select the basis vector, can be one of "random", "mean" or "weighted"}

\item{...}{other parameters (unused, included for compatibility with
generic call)}
}
\value{
Matrix \code{X}' containing the mutated population
}
\description{
Differential Mutation implementation for the MOEA/D
}
\details{
This function generalizes many variations of the Differential Mutation
operator with general form:

u <- x_basis + Phi(x_a - x_b)

Where u is the new candidate vector, Phi is a real number != 0,
and x_basis, x_a and x_b are distinct vectors in the population.

This function include the following variations of this operator:

\itemize{
   \item Phi may be a constant, user provided parameter, or randomly choosen.
   \item X_basis can be a random vector from the population
   \item X_basis can be the mean point in the T neighborhood
   \item X_basis can be the weighted mean point in the T neighborhood
}
}

