% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variation_localsearch.R
\name{variation_localsearch}
\alias{variation_localsearch}
\title{Local search Operators}
\usage{
variation_localsearch(Xt, Yt, Vt, bigZ, sel.indx, B, W, type, which.x, trunc.x,
  ...)
}
\arguments{
\item{Xt}{Matrix of incumbent solutions}

\item{Yt}{Matrix of objective function values for the incumbent solutions
(generated in \code{\link[=evaluate_population]{evaluate_population()}})}

\item{Vt}{List containing information regarding the constraint function
values for the incumbent solutions (generated in \code{\link[=evaluate_population]{evaluate_population()}})}

\item{bigZ}{Matrix containing the scalarized function values for the
incumbent solutions, generated by the call to \code{\link[=scalarize_values]{scalarize_values()}} within
\code{\link[=update_population]{update_population()}}}

\item{B}{Matrix of neighborhoods (generated by \code{\link[=define_neighborhood]{define_neighborhood()}})}

\item{W}{matrix of weights (created by \code{\link[=generate_weights]{generate_weights()}}).}

\item{type}{type of local search. The list of available local search methods
can be generated using \code{get_localsearch_methods()}}

\item{which.x}{logical vector indicating which subproblems should undergo
local search}

\item{trunc.x}{logical vector indicating whether results should be
truncated to the region defined by the box constraints}

\item{...}{other parameters (unused, included for compatibility with
generic call)

B. Chen, W. Zeng, Y. Lin, D. Zhang,
"A new local search-based multiobjective optimization algorithm",
IEEE Trans. Evolutionary Computation 19(1):50-73, 2015.}

\item{bigZ}{Matrix containing the selection indices for each neighborhood,
generated by the call to \code{\link[=order_neighborhood]{order_neighborhood()}} within \code{\link[=update_population]{update_population()}}}
}
\value{
Matrix \code{Xls} containing the modified points (points that did not
undergo local search are indicated as NA in this output matrix).
}
\description{
Local search operators for the MOEA/D
}
\details{
This routine implements the local search operator for the MOEADr package.
This operator is called as any other variation operator, but it presents
some specific features that affect the general behavior of .
\itemize{
\item If the input parameter \code{tau.ls} is present, then this operator is
executed once every \code{tau.ls} iterations.
\item If the input parameter \code{gamma.ls} is present, then this operator is
executed at iterations when the condition \code{runif() < gamma.ls} is
verified.
}

The trigger conditions described above (which are not mutually exclusive) are
verified independently for every subproblem of the MOEA/D, at each iteration.
Whenever the local search is triggered for a given subproblem, it cancels
all other variation operators \emph{for that subproblem} and is executed directly
on the incumbent solution.
}
