% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variation_none.R
\name{variation_none}
\alias{variation_none}
\title{Identity operator}
\usage{
variation_none(X, ...)
}
\arguments{
\item{X}{Population matrix}

\item{...}{other parameters (unused, included for compatibility with
generic call)}
}
\value{
Matrix \code{X}
}
\description{
Identity operator (no variation performed)
}
\details{
Performs the identity operator (no variation). This routine is included to
simplify the use of automated tuning / design tools such as Iterated Racing.
}
