% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variation_polymut.R
\name{variation_polymut}
\alias{variation_polymut}
\title{Polynomial mutation}
\usage{
variation_polymut(X, etam, pm, ...)
}
\arguments{
\item{X}{Population matrix}

\item{etam}{mutation constant}

\item{pm}{variable-wise probability of mutation}

\item{...}{other parameters (unused, included for compatibility with
generic call)}
}
\value{
Matrix \code{X}' containing the mutated population
}
\description{
Polynomial mutation implementation for the MOEA/D
}
\details{
This R implementation of the Polynomial Mutation reproduces the C code
implementation available in package "emoa", version 0.5-0, by Olaf Mersmann.
The differences between the present version and the original one are:
\itemize{
   \item All variables are considered bounded in the (0, 1) interval, which
         simplifies the calculations.
   \item Calculations are vectorized over variables, which also simplifies
         the operator.
}
}
\section{References}{

K. Deb and S. Agrawal (1999). A Niched-Penalty Approach for Constraint
Handling in Genetic Algorithms. In: Artificial Neural Nets and Genetic
Algorithms, pp. 235-243, Springer.

Olaf Mersmann (2012). emoa: Evolutionary Multiobjective
Optimization Algorithms. R package version 0.5-0.\cr
http://CRAN.R-project.org/package=emoa
}

