% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_stop_criteria.R
\name{check_stop_criteria}
\alias{check_stop_criteria}
\title{Stop criteria for MOEA/D}
\usage{
check_stop_criteria()
}
\value{
The function does not return any value. However, it modifies the
state of the calling environment, updating the iteration flag
\code{keep.running}.
}
\description{
Verifies stop criteria for the MOEADr package
}
\section{Parameters}{

This routine accesses all variables defined in the calling environment using
\code{parent.frame()}, so it does not require any explicit input parameters.

However, the calling environment must containat least a variable named
\code{stopcrit}, which is a list vector containing the stop criteria to be
used. Each element of this list vector must have at least a
parameter \code{stopcrit$name}, containing the name of the
criterion to be used, as well any other criterion-specific values
required. The list of available variation operators can be
generated using \code{get_stop_criteria()}. Other variables may also be
required (e.g., \code{iter} counter for iterations or \code{nfe} counter for
number of candidate solutions evaluated), depending on the criteria used.
}

