% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_population.R
\name{create_population}
\alias{create_population}
\title{Create population}
\usage{
create_population(N, problem = NULL)
}
\arguments{
\item{N}{population size}

\item{problem}{list of named problem parameters. See \code{\link{moead}} for
details. If NULL the function searches for \code{problem} in the calling
environment.}
}
\value{
A population matrix X for the MOEA/D.
}
\description{
Create a population for the MOEADr package
}
\details{
This routine creates a population matrix for the MOEA/D. Currently only a
multivariate uniform distribution is implemented. The points are created
within the space 0 <= x_i <= 1, i = 1,...,n_v.
}
