% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_population.R
\name{create_population}
\alias{create_population}
\title{Create population}
\usage{
create_population(N, problem)
}
\arguments{
\item{N}{population size}

\item{problem}{list of named problem parameters. See Section
\code{Problem Description} of the \code{\link[=moead]{moead()}} documentation for details.}
}
\value{
A population matrix X for the MOEA/D.
}
\description{
Create a population for the MOEADr package
}
\details{
This routine creates a population matrix for the MOEA/D. Currently only a
multivariate uniform distribution is implemented. All points are created
within the standardized space \eqn{0 \le x_i \le 1, i = 1,...,n_v}.
}
