% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_population.R
\name{evaluate_population}
\alias{evaluate_population}
\title{Evaluate population}
\usage{
evaluate_population(X, problem)
}
\arguments{
\item{X}{Population matrix of the MOEA/D (each row is a candidate solution).}

\item{problem}{list of named problem parameters. See Section
\code{Problem Description} of the \code{\link[=moead]{moead()}} documentation for details.}
}
\value{
\code{[N x m]} matrix of objective function values, where \code{N} is the
population size and \code{m} is the number of objectives.
}
\description{
Evaluate a population matrix on the objective functions for the MOEADr
package
}
\details{
This routine evaluates a population matrix for the MOEA/D. Each row of the
matrix is considered as a candidate solution. This routine expects the
candidate solutions to be standardized, i.e., that the variable limits given
in \code{problem$xmin} and \code{problem$xmax} are mapped to \code{0} and
\code{1}, respectively.

\strong{Warning}: The calling environment must have a counter variable \code{nfe},
which tracks the total number of function evaluations. This routine
accesses and updates that \code{nfe} variable.
}
