% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_variation.R
\name{perform_variation}
\alias{perform_variation}
\title{Run variation operators}
\usage{
perform_variation(X = NULL, P = NULL, B = NULL, W = NULL,
  variation = NULL, ...)
}
\arguments{
\item{X}{Population matrix of the MOEA/D (each row is a candidate solution).
If \code{NULL} the function searches for \code{X} in the calling environment.}

\item{P}{Matrix of probabilities of selection for variation (created by
\code{\link[=define_neighborhoods]{define_neighborhoods()}}). If \code{NULL} the function searches for
\code{P} in the calling environment.}

\item{B}{Matrix of neighborhood indexes (created by
\code{\link[=define_neighborhoods]{define_neighborhoods()}}). If \code{NULL} the function searches for
\code{B} in the calling environment.}

\item{W}{matrix of weights (created by \code{\link[=generate_weights]{generate_weights()}}). If \code{NULL}
the function searches for \code{W} in the calling environment.}

\item{variation}{List vector containing the variation operators to be used.
See \code{\link[=moead]{moead()}} for details. If \code{NULL} the function searches
for \code{variation} in the calling environment.}

\item{...}{other parameters (unused, included for compatibility with
generic call)}
}
\value{
Modified population matrix X
}
\description{
Sequentially apply variation operators for the MOEADr package
}
\details{
This routine performs the variation block for the MOEA/D. The
list of available variation operators can be generated using
\code{get_variation_operators()}.
}
