% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_moeadoutput.R
\name{plot.moeadoutput}
\alias{plot.moeadoutput}
\title{plot.moeadoutput}
\usage{
\method{plot}{moeadoutput}(moead.output, useArchive = FALSE,
  feasible.only = TRUE, viol.threshold = 1e-06, nondominated.only = TRUE,
  plot.weights = FALSE, which.objectives = NULL, color.by.obj = 1)
}
\arguments{
\item{moead.output}{list object of class \emph{moeadoutput}
(generated by \code{\link[=moead]{moead()}})}

\item{useArchive}{logical flag to use information from \code{moead.output$Archive}.
Only used if moead.output$Archive is not \code{NULL}.}

\item{feasible.only}{logical flag to use only feasible points in the plots.}

\item{viol.threshold}{threshold of tolerated constraint violation, used to
determine feasibility if \code{feasible.only == TRUE}.}

\item{nondominated.only}{logical flag to use only nondominated points in the
plots.}

\item{plot.weights}{logical flag to plot the weight vectors for 2 and
3-objective problems.}

\item{which.objectives}{integer vector of which objectives to plot.
Defaults to \code{NULL} (use all objectives)}

\item{color.by.obj}{integer, determines which objective is used as the basis
for coloring the parallel coordinates plot.}
}
\description{
S3 method for plotting \emph{moeadoutput} objects (the output of \code{\link[=moead]{moead()}}).
}
