% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_moeadoutput.R
\name{plot.moeadoutput}
\alias{plot.moeadoutput}
\title{plot.moeadoutput}
\usage{
\method{plot}{moeadoutput}(moead.output, useArchive = FALSE,
  feasible.only = TRUE, nondominated.only = TRUE, plot.weights = FALSE,
  which.objectives = NULL, ...)
}
\description{
S3 method for plotting \emph{moeadoutput} objects (the output of \code{\link[=moead]{moead()}}).
}
\details{
This function plots the pareto front of the solution set found by moead.
By default, the first and second objectives are plotted, but other objectives
can be specified as parameters.
}
