% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_moeadoutput.R
\name{plot.moeadoutput}
\alias{plot.moeadoutput}
\title{plot.moeadoutput}
\usage{
\method{plot}{moeadoutput}(moead.output, useArchive = FALSE,
  feasible.only = TRUE, viol.threshold = 1e-06, nondominated.only = TRUE,
  plot.weights = FALSE, which.objectives = NULL, color.by.obj = 1, ...)
}
\description{
S3 method for plotting \emph{moeadoutput} objects (the output of \code{\link[=moead]{moead()}}).
}
