% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preset_moead.R
\name{preset_moead}
\alias{preset_moead}
\title{preset_moead}
\usage{
preset_moead(name = NULL)
}
\description{
Standard MOEA/D Preset (Zhang and Li 2007)
}
\details{
Returns a configuration object that replicates the
MOEA/D as in Zhang and Li (2007) (sec. V-E, p.721-722)
}
\examples{
configuration <- preset_moead(name = "original")
configuration$stopcrit = list(list(name = "maxiter", maxiter = 50))

library(smoof)
ZDT1 <- make_vectorized_smoof(prob.name  = "ZDT1",
                              dimensions = 30)
                              problem   <- list(name       = "ZDT1",
                                                xmin       = rep(0, 30),
                                                xmax       = rep(1, 30),
                                                m          = 2)
showpars  <- list(show.iters = "dots", showevery  = 10)
seed      <- 42

output <- moead(problem = problem,
                preset = configuration,
                showpars = showpars, seed = seed)

}
