% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preset_zhang2007.R
\name{preset_moead}
\alias{preset_moead}
\title{preset_moead}
\usage{
preset_moead()
}
\description{
Standard MOEA/D Preset (Zhang and Li 2007)
}
\details{
Returns a configuration object that replicates the
MOEA/D as in Zhang and Li (2007) (sec. V-E, p.721-722)
}
