% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_progress.R
\name{print_progress}
\alias{print_progress}
\title{Print progress of MOEA/D}
\usage{
print_progress(iter.times, showpars)
}
\arguments{
\item{iter.times}{vector of iteration times of the \code{\link[=moead]{moead()}} routine.}

\item{showpars}{list object containing parameters that control the printed
output of \code{\link[=moead]{moead()}}. Parameter \code{showpars} can have the
following key-value pairs:
\itemize{
\item \verb{$show.iters}: type of output ("dots", "numbers", or "none"). If
not present in \code{showpars}, it defaults to "numbers";
\item \verb{$showevery}: positive integer that determines how frequently
the routine echoes something to the terminal. If not present in
\code{showpars}, it defaults to \code{10}.
}}
}
\description{
Echoes progress of MOEA/D to the terminal for the MOEADr package
}
\section{References}{

F. Campelo, L.S. Batista, C. Aranha (2020): The {MOEADr} Package: A
Component-Based Framework for Multiobjective Evolutionary Algorithms Based on
Decomposition. Journal of Statistical Software \doi{10.18637/jss.v092.i06}\cr
}

