% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalarize_values.R
\name{scalarize_values}
\alias{scalarize_values}
\title{Scalarize values for MOEA/D}
\usage{
scalarize_values(moead.env, normYs, B)
}
\arguments{
\item{moead.env}{list representing the environment of the base function
\code{moead}.}

\item{normYs}{list containing matrices of scaled objective values (see
section \code{Parameters}).}

\item{B}{neighborhood matrix, generate by \code{\link{define_neighborhood}}.}
}
\value{
\code{[ (T+1) x N ]} matrix of scalarized performance values. Each column
contains the T scalarized performances of the candidate solutions in the
neighborhood of a given subproblem, plus the scalarized performance value
for the incumbent solution for that subproblem.
}
\description{
Perform scalarization for the MOEADr package.
}
\details{
This routine calculates the scalarized performance values for the MOEA/D.
}
\section{Parameters}{

This routine receives two input variables:
\itemize{
\item \code{moead.env} is generated within the calling function
\code{update_population()}. See \code{\link{update_population}} for more
information.
\item\code{normYs} is a list generated by function
\code{scale_objectives(...)}, containing two matrices of scaled objective
values (\code{Y} and \code{Yt}) and two vectors, containing the current
estimates of the ideal (\code{minP}) and nadir (\code{maxP}) points. See
\code{\link{scale_objectives}} for details.
}

The environment listed in \code{moead.env} must contain a variable
\code{aggfun}, which provides the type of scalar aggregation function to be
employed. The list of available scalarization methods can be generated using
\code{get_scalarization_methods()}
}

