% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_moeadoutput.R
\name{summary.moeadoutput}
\alias{summary.moeadoutput}
\title{summary.moeadoutput}
\usage{
\method{summary}{moeadoutput}(moead.output, useArchive = FALSE,
  viol.threshold = 1e-06, ndigits = 3, ...)
}
\arguments{
\item{moead.output}{list object of class \emph{moeadoutput}
(generated by \code{\link[=moead]{moead()}})}

\item{useArchive}{logical flag to use information from \code{moead.output$Archive}.
Only used if moead.output$Archive is not \code{NULL}.}

\item{viol.threshold}{threshold of tolerated constraint violation, used to
determine feasibility of points in \code{moead.output}.}

\item{ndigits}{number of decimal places to use for the ideal and nadir estimates}
}
\description{
S3 method for summarizing \emph{moeadoutput} objects (the output of \code{\link[=moead]{moead()}}).
}
