% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_moeadoutput.R
\name{summary.moeadoutput}
\alias{summary.moeadoutput}
\title{summary.moeadoutput}
\usage{
\method{summary}{moeadoutput}(object, ..., useArchive = FALSE,
  viol.threshold = 1e-06, ndigits = 3)
}
\arguments{
\item{object}{list object of class \emph{moeadoutput}
(generated by \code{\link[=moead]{moead()}})}

\item{...}{other parameters to be passed down to specific summary functions (currently unused)}

\item{useArchive}{logical flag to use information from \code{object$Archive}.
Only used if object$Archive is not \code{NULL}.}

\item{viol.threshold}{threshold of tolerated constraint violation, used to
determine feasibility of points in \code{object}.}

\item{ndigits}{number of decimal places to use for the ideal and nadir estimates}
}
\description{
S3 method for summarizing \emph{moeadoutput} objects (the output of \code{\link[=moead]{moead()}}).
}
