% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updt_best.R
\name{updt_best}
\alias{updt_best}
\title{"Best" Neighborhood Replacement Update for MOEA/D}
\usage{
updt_best(moead.env)
}
\arguments{
\item{moead.env}{list representing the environment of the base function
\code{moead}.}

\item{nr}{A positive integer. It determinines the maximum number of times that
any individual can be selected (if nr is equal to the population size,
updated_restricted behaves exactly as updated_standard)}

\item{Tr}{A positive integer. It determines the maximum size of the neighborhood
calculated by the Best Neighborhood method.}
}
\value{
List object containing the update population matrix (\code{Xnext})
and its corresponding matrix of objective function values (\code{Ynext}).
}
\description{
Population update using the "best neighborhood" replacement
method for the MOEADr package.
}
\details{
The Best Neighborhood replacement method resolves in three
steps:

1- For each subproblem i, the best individual x_j from the
   entire population is chosen.

2- The neighborhood of subproblem i is replaced by the
   neighborhood of problem j. The size of this neighborhood
   is truncated by parameter \code{Tr}.

3- The Restricted replacement method is applied to this
   new neighborhood.
}
\section{Parameters}{

This routine receives a single input variable, \code{moead.env}, which is
generated within the calling function \code{update_population()}. See
\code{\link{update_population}} for more information.
}

