% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updt_restricted.R
\name{updt_restricted}
\alias{updt_restricted}
\title{Restricted Neighborhood Replacement Update for MOEA/D}
\usage{
updt_restricted(moead.env)
}
\arguments{
\item{moead.env}{list representing the environment of the base function
\code{moead}.

This routine also expects the list \code{moead.env$updated} to be populated
with the following names:}

\item{nr}{A positive integer. It determines the maximum number of times that
any individual can be selected (if nr is equal to the population size,
updated_restricted behaves exactly as updated_standard)}
}
\value{
List object containing the update population matrix (\code{Xnext})
and its corresponding matrix of objective function values (\code{Ynext}).
}
\description{
Population update using the restricted neighborhood replacement
method for the MOEADr package.
}
\details{
The restricted neighborhood replacement method behaves like
the "standard" replacement method, except that each individual
can only be selected up to \code{nr} times. After this limit
has been reached, the next best individual in the same
neighborhood is selected.
}
\section{Parameters}{


This routine receives a single input variable, \code{moead.env}, which is
generated within the calling function \code{update_population()}. See
\code{\link{update_population}} for more information.
}

