% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updt_standard.R
\name{updt_standard}
\alias{updt_standard}
\title{Standard Neighborhood Replacement Update for MOEA/D}
\usage{
updt_standard(moead.env)
}
\arguments{
\item{moead.env}{list representing the environment of the base function
\code{moead}.}
}
\value{
List object containing the update population matrix (\code{Xnext})
and its corresponding matrix of objective function values (\code{Ynext}).
}
\description{
Population update using the standard neighborhood replacement method for the
MOEADr package.
}
\details{
This routine executes the standard neighborhood replacement operation to
update the population matrix of the MOEA/D.
}
\section{Parameters}{

This routine receives a single input variable, \code{moead.env}, which is
generated within the calling function \code{\link[=update_population]{update_population()}}. See
the documentation of that function for more information.
}

