% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variation_diffmut.R
\name{variation_diffmut}
\alias{variation_diffmut}
\title{Differential Mutation}
\usage{
variation_diffmut(X, P, B, Phi, basis = "rand", ...)
}
\arguments{
\item{X}{Population matrix}

\item{P}{Matrix of selection probabilities (generated by
\code{define_neighborhood(...)})}

\item{B}{Matrix of neighborhoods (generated by
\code{define_neighborhood(...)})}

\item{Phi}{Mutation parameter, random if NULL}

\item{basis}{how to select the basis vector, can be one of "random", "mean" or "wgi"}

\item{...}{other parameters (unused, included for compatibility with
generic call)}
}
\value{
Matrix \code{X}' containing the mutated population
}
\description{
Differential Mutation implementation for the MOEA/D
}
\details{
This function generalizes many variations of the Differential Mutation
operator with general form:

u <- x_basis + Phi(x_a - x_b)

Where u is the new candidate vector, Phi is a real number != 0,
and x_basis, x_a and x_b are distinct vectors from the population.

This function include the following variations of this operator:

\itemize{
\item Phi may be either constant (user provided) or randomly chosen.
\item x_basis can be a random vector from the population
(\code{basis = "rand"})
\item x_basis can be the mean point in the T neighborhood points
(\code{basis = "mean"})
\item x_basis can be the weighted mean point in the T neighborhood points
(\code{basis = "wgi"})
}
}
\section{References}{

K. Price, R.M. Storn, J.A. Lampinen, "Differential Evolution: A
Practical Approach to Global Optimization", Springer 2005

D. V. Arnold, “Weighted multirecombination evolution strategies,”
Theoretical Computer Science 361(1):18–37, 2006.
}

