% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variation_localsearch.R
\name{variation_localsearch}
\alias{variation_localsearch}
\title{Local search Operators}
\usage{
variation_localsearch(X, Xt, Yt, B, type, which.x, trunc.x, ...)
}
\arguments{
\item{X}{Matrix of candidate solutions (generated in the calling function
\code{\link[=perform_variation]{perform_variation()}})}

\item{Xt}{Matrix of incumbent solutions}

\item{Yt}{Matrix of objective function values for the incumbent solutions}

\item{B}{Matrix of neighborhoods (generated by \code{\link[=define_neighborhood]{define_neighborhood()}})}

\item{type}{type of local search. The list of available local search methods
can be generated using \code{get_localsearch_methods()}}

\item{which.x}{logical vector indicating which subproblems should undergo
local search}

\item{trunc.x}{logical vector indicating whether results should be
truncated to the region defined by the box constraints}

\item{...}{other parameters (unused, included for compatibility with
generic call)}
}
\value{
Matrix \code{X}' containing the modified population
}
\description{
Local search operators for the MOEA/D
}
\details{
This routine implements the local search operator for the MOEADr package.
This operator is called as any other variation operator, but it presents
some specific features that affect the general behavior of .
\itemize{
\item If the input parameter \code{tau.ls} is present, then this operator is
executed once every \doce{tau.ls} iterations.
\item If the input parameter \code{gamma.ls} is present, then this operator is
executed at iterations when the condition \code{runif() < gamma.ls} is
verified.
}

The trigger conditions described above (which are not mutually exclusive) are
verified independently for every subproblem of the MOEA/D, at each iteration.
Whenever the local search is triggered for a given subproblem, it cancels
all other variation operators \emph{for that subproblem} and is executed directly
on the incumbent solution.
}
\section{References}{

Y. Tan, Y. Jiao, H. Li, X. Wang,
"A modification to MOEA/D-DE for multiobjective optimization problems with
complicated Pareto sets",
Information Sciences 213(1):14-38, 2012.

B. Chen, W. Zeng, Y. Lin, D. Zhang,
"A new local search-based multiobjective optimization algorithm",
IEEE Trans. Evolutionary Computation 19(1):50-73, 2015.
}

