% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variation_truncate.R
\name{variation_truncate}
\alias{variation_truncate}
\title{Truncate}
\usage{
variation_truncate(X, ...)
}
\arguments{
\item{X}{Population matrix}

\item{...}{other parameters (unused, included for compatibility with
generic call)}
}
\value{
Matrix \code{X}
}
\description{
Truncation variation operator
}
\details{
Truncate the solution matrix X to guarantee that all values are in the
[0, 1] interval.
}
\section{References}{

Code reused from https://github.com/fcampelo/ExpDE/blob/master/R/ExpDE.R
Lines 240-242
}

