% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variation_truncate.R
\name{variation_truncate}
\alias{variation_truncate}
\title{Truncate}
\usage{
variation_truncate(X, ...)
}
\arguments{
\item{X}{Population matrix}

\item{...}{other parameters (included for compatibility with generic call)}
}
\value{
Truncated matrix \code{X}'.
}
\description{
Truncation variation operator
}
\details{
Truncate the solution matrix \code{X} to the \verb{[0, 1]} interval.
}
\section{References}{


F. Campelo, L.S. Batista, C. Aranha (2020): The {MOEADr} Package: A
Component-Based Framework for Multiobjective Evolutionary Algorithms Based on
Decomposition. Journal of Statistical Software \doi{10.18637/jss.v092.i06}\cr
}

